Create charPicGrid.py
Say you have a list of lists where each value in the inner lists is a one-character string. Think of grid[x][y] as being the character at the x- and y-coordinates of a “picture” drawn with text characters. The (0, 0) origin is in the upper-left corner, the x-coordinates increase going right, and the y-coordinates increase going down. Write code to rotate the image 90° clockwise